# backend/openapi/v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "sCalenderPlus API",
    "description": "Public REST API of sCalenderPlus. See docs/architecture/api.md.",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Create an account",
        "description": "Always answers 202 for a well-formed request with an allowed email domain (disposable-email providers: 422 email_domain_not_allowed): a new account gets a confirmation email, an already registered address gets a hint email instead (no user enumeration). Sign in afterwards with POST /auth/login.",
        "operationId": "Register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Accepted"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "default": {
            "description": "Error (RFC 9457 problem details with a stable `code`).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Sign in with email and password",
        "description": "Sets the session cookie, or answers twoFactorRequired (second step: POST /auth/login/2fa). Failures are 401 invalid_credentials; after 5 failures in a row the account is locked for 15 minutes (still answered as invalid_credentials).",
        "operationId": "Login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error (RFC 9457 problem details with a stable `code`).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/login/2fa": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Second login step: authenticator code or recovery code",
        "description": "Needs the pending-login cookie from POST /auth/login (twoFactorRequired, valid 5 minutes); without it 401 unauthenticated. Wrong codes are 401 invalid_credentials and count towards the lockout; each recovery code works once.",
        "operationId": "LoginTwoFactor",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginTwoFactorRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "default": {
            "description": "Error (RFC 9457 problem details with a stable `code`).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "End the session (clears the session cookie)",
        "operationId": "Logout",
        "responses": {
          "204": {
            "description": "No Content"
          },
          "default": {
            "description": "Error (RFC 9457 problem details with a stable `code`).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": [ ]
          }
        ]
      }
    },
    "/api/v1/auth/confirm-email": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Confirm an email address with the token from the confirmation link",
        "operationId": "ConfirmEmail",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmEmailRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "default": {
            "description": "Error (RFC 9457 problem details with a stable `code`).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/confirm-email/resend": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Send the confirmation email again to the signed-in user (no-op when already confirmed)",
        "operationId": "ResendEmailConfirmation",
        "responses": {
          "204": {
            "description": "No Content"
          },
          "default": {
            "description": "Error (RFC 9457 problem details with a stable `code`).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": [ ]
          }
        ]
      }
    },
    "/api/v1/auth/forgot-password": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Request a password reset email",
        "description": "Always answers 202 (no user enumeration); only registered addresses receive an email.",
        "operationId": "ForgotPassword",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Accepted"
          },
          "default": {
            "description": "Error (RFC 9457 problem details with a stable `code`).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/reset-password": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Set a new password with the token from the reset link",
        "description": "Also confirms the email address (the link proved ownership), lifts a lockout and ends all other sessions.",
        "operationId": "ResetPassword",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "default": {
            "description": "Error (RFC 9457 problem details with a stable `code`).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ConfirmEmailRequest": {
        "required": [
          "userId",
          "token"
        ],
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "token": {
            "maxLength": 2048,
            "type": "string"
          }
        }
      },
...
      "ErrorCode": {
        "enum": [
          "bad_request",
          "calendar_frozen",
          "conflict",
          "csrf_header_missing",
          "email_domain_not_allowed",
          "email_not_verified",
          "external_sharing_not_allowed",
          "feature_not_in_plan",
          "insufficient_permission",
          "internal_error",
          "invalid_credentials",
          "method_not_allowed",
          "not_found",
          "override_invalid_in_target",
          "payload_too_large",
          "permission_self_lockout",
          "plan_limit_reached",
          "precondition_failed",
          "precondition_required",
          "rate_limited",
          "reauthentication_failed",
          "recurrence_invalid",
          "service_unavailable",
          "time_zone_invalid",
          "token_expired",
          "token_invalid",
          "two_factor_required",
          "uid_conflict",
          "unauthenticated",
          "unsupported_media_type",
          "validation_failed"
        ],
        "type": "string",
        "description": "Stable machine-readable error code (part of the contract; never renamed)."
      },
      "ForgotPasswordRequest": {
        "required": [
          "email"
        ],
        "type": "object",
        "properties": {
          "email": {
            "maxLength": 256,
            "type": "string"
          }
        }
      },
...
      "HttpValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
      "LoginRequest": {
        "required": [
          "email",
          "password"
        ],
        "type": "object",
        "properties": {
          "email": {
            "maxLength": 256,
            "type": "string"
          },
          "password": {
            "maxLength": 128,
            "type": "string"
          },
          "rememberMe": {
            "type": "boolean"
          }
        }
      },
      "LoginResponse": {
        "required": [
          "twoFactorRequired",
          "user"
        ],
        "type": "object",
        "properties": {
          "twoFactorRequired": {
            "type": "boolean"
          },
          "user": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/MeResponse"
              }
            ]
          }
        }
      },
      "LoginTwoFactorRequest": {
        "type": "object",
        "properties": {
          "code": {
            "maxLength": 16,
            "minLength": 0,
            "type": [
              "null",
              "string"
            ]
          },
          "recoveryCode": {
            "maxLength": 32,
            "minLength": 0,
            "type": [
              "null",
              "string"
            ]
          },
          "rememberMe": {
            "type": "boolean"
          }
        }
      },
...
      "ProblemDetails": {
        "required": [
          "type",
          "title",
          "status",
          "code"
        ],
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "format": "uri"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          },
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "traceId": {
            "type": "string"
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "description": "Validation errors by field (code `validation_failed`)."
          }
        },
        "description": "RFC 9457 problem details. Further members depend on the code (e.g. `limit`, `required`, `actual`)."
      },
...
      "RegisterRequest": {
        "required": [
          "email",
          "password",
          "displayName"
        ],
        "type": "object",
        "properties": {
          "email": {
            "maxLength": 256,
            "type": "string"
          },
          "password": {
            "maxLength": 128,
            "minLength": 0,
            "type": "string"
          },
          "displayName": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          },
          "locale": {
            "type": [
              "null",
              "string"
            ]
          },
          "timeZone": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "ResetPasswordRequest": {
        "required": [
          "userId",
          "token",
          "newPassword"
        ],
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "token": {
            "maxLength": 2048,
            "type": "string"
          },
          "newPassword": {
            "maxLength": 128,
            "minLength": 0,
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "description": "Session cookie set by POST /api/v1/auth/login (web app, same origin). Unsafe methods also need the header `X-Requested-With: scal`.",
        "name": "__Host-scal",
        "in": "cookie"
      }
    }
  },
...
}